+++ src/drivers/i2c.rs
@@ -15,14 +15,14 @@ static BUSES: LazyLock<SpinLocked<Vec<Result<hal::i2c::Bus>, 3>>> = LazyLock::ne
         let bus = hal::i2c::init(cfg).map_err(|e| e.into());
 
         match &bus {
-            Ok(bus) => kprint!("    Initialized I2C bus at 0x{:x}\n", cfg.instance),
+            Ok(_) => kprint!("    Initialized I2C bus at 0x{:x}\n", cfg.instance),
             Err(e) => kprint!(
                 "    Failed to initialize I2C bus at 0x{:x}: {e}\n",
                 cfg.instance
             ),
         }
 
-        buses.push(bus);
+        buses.push(bus).expect("Bus must fit in inline storage.");
     }
     SpinLocked::new(buses)
 });
@@ -61,25 +61,37 @@ impl Device {
         Err(kerr!(EINVAL))
     }
 
-    pub fn write(&self, tx: &[u8]) -> Result<()> {
-        match hal::i2c::write(&self.desc, tx) {
+    pub fn write(&self, tx: &[u8], timeout: u16) -> Result<()> {
+        match hal::i2c::write(&self.desc, tx, timeout) {
             Ok(()) => Ok(()),
             Err(e) => Err(e.into()),
         }
     }
 
-    pub fn read(&self, rx: &mut [u8]) -> Result<()> {
-        match hal::i2c::read(&self.desc, rx) {
+    pub fn read(&self, rx: &mut [u8], timeout: u16) -> Result<()> {
+        match hal::i2c::read(&self.desc, rx, timeout) {
             Ok(()) => Ok(()),
             Err(e) => Err(e.into()),
         }
     }
 
-    pub fn write_read(&self, tx: &[u8], rx: &mut [u8]) -> Result<()> {
-        match hal::i2c::write_read(&self.desc, tx, rx) {
+    pub fn write_read(&self, tx: &[u8], rx: &mut [u8], timeout: u16) -> Result<()> {
+        match hal::i2c::write_read(&self.desc, tx, rx, timeout) {
             Ok(()) => Ok(()),
             Err(e) => Err(e.into()),
         }
+    }
+
+    pub fn check_and_recover_bus(&self) -> Result<bool> {
+        match hal::i2c::bus_recovery_needed(&self.desc) {
+            Ok(true) => match hal::i2c::recover_bus(&self.desc) {
+                Ok(()) => return Ok(true),
+                Err(e) => return Err(e.into()),
+            },
+            Ok(false) => {}
+            Err(e) => return Err(e.into()),
+        }
+        Ok(false)
     }
 }
