+++ src/drivers/spi.rs
@@ -14,13 +14,13 @@ static BUSES: LazyLock<SpinLocked<Vec<Result<hal::spi::Bus>, 3>>> = LazyLock::ne
     for cfg in hal::device_tree::SPI_BUS_REGISTRY {
         let bus = hal::spi::init(cfg).map_err(|e| e.into());
         match &bus {
-            Ok(bus) => kprint!("    Initialized SPI bus at 0x{:x}\n", cfg.instance),
+            Ok(_) => kprint!("    Initialized SPI bus at 0x{:x}\n", cfg.instance),
             Err(e) => kprint!(
                 "    Failed to initialize SPI bus at 0x{:x}: {e}\n",
                 cfg.instance
             ),
         }
-        buses.push(bus);
+        buses.push(bus).expect("Bus must fit in inline storage.");
     }
     SpinLocked::new(buses)
 });
