+++ src/lib.rs
@@ -28,7 +28,7 @@ pub use hal_cortex_m::*;
 // Add new hals here. No cfg needed.
 
 pub use hal::Machinelike;
-pub use hal_api::PosixError;
+pub use hal_api::error::*;
 pub use proc_macros::app_main;
 
 /// The kernel initialization function.
