+++ src/types.rs
@@ -4,6 +4,7 @@ pub mod boxed;
 pub mod heap;
 pub mod list;
 pub mod pool;
+pub mod queue;
 pub mod rbtree;
 pub mod traits;
 pub mod view;
